성적 산출 기준 변경 중...

+++ src/main/java/com/green/university/entity/StuSub.java
@@ -23,5 +23,5 @@ public class StuSub { // 학생의 수강 과목과 학점
     @JoinColumn(name = "grade_id")
     private Grade grade;
 
-    private Long completeGrade; // 환산점수
+    private double completeGrade; // 환산점수
 }
